do not fetch token until someone hits submit button
