+++ static/js/src/elements/ManualPay.vue
@@ -3,10 +3,9 @@
     :class="classesWithValidation"
   >
     <card
-      ref="card"
       :options="options"
       :stripe="stripeKey"
-      @change="onChange($event.complete)"
+      @change="onChange"
     />
     <p
       v-if="showError && !valid"
@@ -18,11 +17,9 @@
 </template>
 
 <script>
-import { Card, createToken } from 'vue-stripe-elements-plus';
+import { Card } from 'vue-stripe-elements-plus';
 
-import updateStoreValue from './mixins/updateStoreValue';
-import getStoreValue from './mixins/getStoreValue';
-import createCustomer from '../utils/createCustomer';
+import updateValidity from './mixins/updateValidity';
 
 export default {
   name: 'ManualPay',
@@ -31,10 +28,7 @@ export default {
     Card,
   },
 
-  mixins: [
-    updateStoreValue,
-    getStoreValue,
-  ],
+  mixins: [updateValidity],
 
   props: {
     errorClasses: {
@@ -45,16 +39,6 @@ export default {
     showError: {
       type: Boolean,
       default: false,
-    },
-
-    customerIdStoreModule: {
-      type: String,
-      required: true,
-    },
-
-    emailStoreModule: {
-      type: String,
-      required: true,
     },
 
     validation: {
@@ -95,91 +79,11 @@ export default {
   },
 
   methods: {
-    markFetchingToken() {
-      this.$emit('setValue', { key: 'isFetchingToken', value: true });
-    },
-
-    markNotFetchingToken() {
-      this.$emit('setValue', { key: 'isFetchingToken', value: false });
-    },
-
-    markValid(element) {
-      this.$emit('setValidationValue', { element, key: 'valid', value: true });
-    },
-
-    markInvalid(element) {
-      this.$emit('setValidationValue', { element, key: 'valid', value: false });
-    },
-
-    markErrorAndInvalid({ element, message }) {
-      const updates = [
-        { element, key: 'valid', value: false },
-        { element, key: 'message', value: message },
-      ];
-
-      this.$emit('setValidationValue', updates);
-    },
-
-    showManualErrors() {
-      this.$emit('setValue', { key: 'showManualErrors', value: true });
-    },
-
-    onChange(isComplete) {
-      if (isComplete) {
-        this.markFetchingToken();
-
-        createToken().then((result) => {
-          if (!result.error) {
-            const { token: { id: token } } = result;
-            const email = this.getStoreValue({
-              storeModule: this.emailStoreModule,
-              key: 'stripeEmail',
-            });
-
-            /**
-             Because Stripe 3 does not validate CVC client side,
-             we have to create the customer on the server and
-             check for errors returned there. If they exist,
-             we display them. If not, we store the returned customer ID.
-            */
-            createCustomer({ token, email })
-              .then(({ data: { customer_id: customerId } }) => {
-                this.updateStoreValue({
-                  storeModule: this.customerIdStoreModule,
-                  key: 'customerId',
-                  value: customerId,
-                });
-                this.markValid('card');
-              })
-              .catch(({ response: { data: { type, message } } }) => {
-                let element;
-
-                if (type === 'email') {
-                  this.$refs.card.clear();
-                  element = 'stripeEmail';
-                  this.showManualErrors();
-                } else if (type === 'card') {
-                  element = 'card';
-                }
-
-                this.markErrorAndInvalid({ element, message });
-              })
-              .then(() => {
-                this.markNotFetchingToken();
-              });
-          } else {
-            this.markNotFetchingToken();
-            this.markErrorAndInvalid({
-              element: 'card',
-              message: result.error.message,
-            });
-          }
-        });
+    onChange({ error }) {
+      if (error) {
+        this.markMessageAndInvalid({ element: 'card', message: error.message });
       } else {
-        this.markErrorAndInvalid({
-          element: 'card',
-          message: 'Invalid card input',
-        });
+        this.markValid('card');
       }
     },
   },
